requirements.txt
django==3.2.7
djangorestframework
django-cors-headers
pillow
django-environ
django-filter
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
psycopg2-binary
drf-yasg
# redis==4.3.1
# celery==5.2.6